fix some warnings

+++ core/src/main/java/com/watabou/pixeldungeon/items/wands/WandOfLightning.java
@@ -67,7 +67,7 @@ public class WandOfLightning extends Wand {
 		}
 		
 		affected.add( ch );
-		ch.damage( Level.water[ch.pos] && !ch.flying ? (int)(damage * 2) : damage, LightningTrap.LIGHTNING  );
+		ch.damage( Level.water[ch.pos] && !ch.flying ? (damage * 2) : damage, LightningTrap.LIGHTNING  );
 		
 		ch.getSprite().centerEmitter().burst( SparkParticle.FACTORY, 3 );
 		ch.getSprite().flash();
